Add files via upload
The Multi Vision Recipe Recommendation App will be very beneficial for the 
users and even for the people who are suffering from multiple diseases. Direct 
benefits of the app will be that it provides a recommendation system on the 
basis of ingredients which are present at that moment. It also allows the user to 
take the images of the ingredients and on the basis of that images the user will 
be provided with recommended recipes. Indirect benefits of the app will in that
sense that it will help the user to save the time and money to use the ingredients 
which he/she can afford and even use the ingredients which are left. It will also 
3
be very beneficial for the patients and the peoples who are suffering from 
different diseases in the sense that it can provide the recommendations of the 
recipe on the basis of ingredients. The people having different allergies from
different ingredients can also take benefit by using this app in that manner that 
they can provide the ingredients which are suitable for them and in return they 
can have recommended recipes. You capture a picture of vegetable and now our application will 
give you the recipes after recognizing that you have entered the picture of a 
fruit. Or you can input the ingredient manually from the application.